Python/MicrosoftFoundry.Classic/requirements.txt
agent-framework --pre
agent-framework-azure-ai --pre
azure-identity==1.25.1
pydantic==2.12.5
python-dotenv